refactor: split matrix operations into separate modules
- Split s21_matrix.c into individual module files (one function per file)
- Each operation now has its own module following single responsibility:
  * s21_create_matrix.c
  * s21_remove_matrix.c
  * s21_eq_matrix.c
  * s21_sum_matrix.c
  * s21_sub_matrix.c
  * s21_mult_number.c
  * s21_mult_matrix.c
  * s21_transpose.c
  * s21_calc_complements.c
  * s21_determinant.c
  * s21_inverse_matrix.c
- s21_matrix.c now serves as entry point with documentation
- Add missing stddef.h includes for NULL definition
- Follow modular programming principles
- All 70 tests passing with no memory leaks
